Allow to perform the single sign-on with an external application
This allows to execute an external application to retrieve the
SVPNCOOKIE, automating the process for a more user friendly process.
The single sign-on likely require the use of a browser.

The application should write the SVPNCOOKIE to stdout and exit.

The option '--saml-handler' relies on 'sudo --preserve-env' to work:
the external application should run as unprivileged user with all the
standard env variables set (this is the case for Chromium-based
applications), but openfortivpn must run as root. Requiring the use
of sudo simply allows to satisfy the above requirements keeping
'openfortivpn' almost unchanged.